Moved typedefs to PlatformDefs
