+++ libraries/platformdefs/CMakeLists.txt
@@ -2,6 +2,7 @@ add_library(${TARGETNAME_LIB_PLATFORMDEFS} INTERFACE
 	include/PlatformDefs/decorators.h
 	include/PlatformDefs/libnames.h
 	include/PlatformDefs/platformid.h
+	include/PlatformDefs/typedefs.h
 )
 
 target_include_directories(${TARGETNAME_LIB_PLATFORMDEFS} INTERFACE include)
