+++ libraries/platformdefs/include/PlatformDefs/typedefs.h
